requirements.txt
streamlit==1.32.0
openpyxl==3.1.5
requests==2.32.3
python-dotenv==1.0.1
